README.md
This program obtains the information of a product and displays it for the user. The information displayed are: who manufactured it,
serial number, date, and name of the product. The Enum class stores all the constant item types such as AUDIO and VISUAL. The Product 
class stores all the information for the products, and main executes the details of those products.